fix: SQLAlchemy sync_engine, plain pytest for observability tests
- telemetry.py: instrument engine.sync_engine — AsyncEngine blocks event.listen();
  sync_engine is the correct target; spans still produced as all async queries
  route through it. Remove is_instrumented_by() (not available on this version).
- Replace BDD observability tests with plain pytest (test_observability.py) —
  span delivery to Jaeger is infrastructure verification, not user-facing behavior;
  plain pytest matches the pattern used by test_db_schema.py and test_integration.py
